+++ src/PredisStore.php
@@ -15,6 +15,7 @@ use Exception;
 use Predis\Client;
 use Predis\ClientInterface;
 use Webmozart\KeyValueStore\Api\KeyValueStore;
+use Webmozart\KeyValueStore\Api\NoSuchKeyException;
 use Webmozart\KeyValueStore\Api\ReadException;
 use Webmozart\KeyValueStore\Api\WriteException;
 use Webmozart\KeyValueStore\Assert\Assert;
@@ -66,18 +67,22 @@ class PredisStore implements KeyValueStore
     /**
      * {@inheritdoc}
      */
-    public function get($key, $default = null)
+    public function get($key)
     {
         KeyUtil::validate($key);
 
-        try {
-            if (!$this->client->exists($key)) {
-                return $default;
-            }
+        $serialized = null;
 
-            $serialized = $this->client->get($key);
+        try {
+            if ($this->client->exists($key)) {
+                $serialized = $this->client->get($key);
+            }
         } catch (Exception $e) {
             throw ReadException::forException($e);
+        }
+
+        if (null === $serialized) {
+            throw NoSuchKeyException::forKey($key);
         }
 
         return Serializer::unserialize($serialized);