+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 * 1.0.0-next (@release_date@)
 
  * added `KeyValueStore::keys()`
+ * made `KeyValueStore::get()` throw an exception when a key is not found to
+   prevent superfluous calls to `has()`
 
 * 1.0.0-beta3 (2015-04-13)